style: getFolderList 함수 api.js 파일로 이동
